Fix parent identifier extension to 64 bits.
Source location identifiers are signed 16 bits. Extending this value to
64 bits without first casting it to unsigned 16 bit caused bit extension
of the sign bit, making the value clash with "unselected" group
identifier.

+++ server/TracyView.cpp
@@ -9187,7 +9187,7 @@ void View::DrawFindZone()
                 const auto parent = GetZoneParent( *ev.Zone(), m_worker.DecompressThread( ev.Thread() ) );
                 if( parent )
                 {
-                    group = &m_findZone.groups[uint64_t( parent->SrcLoc() )];
+                    group = &m_findZone.groups[uint64_t( uint16_t( parent->SrcLoc() ) )];
                 }
                 else
                 {
